[Refactor] MainPage로 함수 이동
